Factored out a common basis for multiple file downloads to be reused in ATOM and WFS paging downloads.
